PHPStan level 7 scan

+++ libraries/user/session/Config.php
@@ -9,11 +9,12 @@ use df;
 use df\core;
 use df\user;
 
-class Config extends core\Config {
-
+class Config extends core\Config
+{
     const ID = 'Session';
 
-    public function getDefaultValues(): array {
+    public function getDefaultValues(): array
+    {
         return [
             'gc' => [
                 'probability' => null
@@ -29,35 +30,40 @@ class Config extends core\Config {
     }
 
 
-// GC
-    public function getGcProbability(): int {
-        return core\math\Util::clampInt(
+    // GC
+    public function getGcProbability(): int
+    {
+        return (int)core\math\Util::clampInt(
             $this->values->gc->get('probability', 3),
             1, 10
         );
     }
 
-// Transitions
-    public function transitionsEnabled(): bool {
+    // Transitions
+    public function transitionsEnabled(): bool
+    {
         return (bool)$this->values->transitions->get('enabled', true);
     }
 
-    public function getTransitionProbability(): int {
-        return core\math\Util::clampInt(
+    public function getTransitionProbability(): int
+    {
+        return (int)core\math\Util::clampInt(
             $this->values->transitions->get('probability', 10),
             2, 15
         );
     }
 
-    public function getTransitionLifeTime(): int {
-        return core\math\Util::clampInt(
+    public function getTransitionLifeTime(): int
+    {
+        return (int)core\math\Util::clampInt(
             $this->values->transitions->get('lifeTime', 30),
             10, 60
         );
     }
 
-    public function getTransitionCooloff(): int {
-        return core\math\Util::clampInt(
+    public function getTransitionCooloff(): int
+    {
+        return (int)core\math\Util::clampInt(
             $this->values->transitions->get('cooloff', 20),
             10, 60
         );
